requirements.txt
torch_scatter
ninja
tqdm
imageio
imageio-ffmpeg
mmcv
taichi==1.2